+++ libs/category/src/models/category-request.ts
@@ -1,7 +1,7 @@
 import { ID } from '@daffodil/core';
 import { DaffSortDirectionEnum } from '@daffodil/core/state';
 
-import { DaffCategoryFilterRequest } from './filter-request';
+import { DaffCategoryFilterRequest } from './requests/filter-request';
 
 export interface DaffCategoryRequest {
   id: ID;
