+++ libs/category/src/models/public_api.ts
@@ -8,7 +8,7 @@ export { DaffCategoryPageConfigurationState } from './category-page-configuratio
 export { DaffCategoryBreadcrumb } from './category-breadcrumb';
 export { DaffCategory } from './category';
 export { DaffGenericCategory } from './generic-category';
-export { DaffCategoryRequest } from './requests/category-request';
+export { DaffCategoryRequest } from './category-request';
 export * from './requests/filter-request';
 export * from './category-filter-base';
 export * from './category-applied-filter';
